requirements.txt
enum==0.4.4
grizzled-python==1.0.1
Jinja2==2.7.3
markdown2==2.3.0
MarkupSafe==0.23
MySQL-python==1.2.5
prettytable==0.7.2
PyYAML==3.11
sae-python-dev==1.3.2
sqlcmd==0.7.1
Werkzeug==0.10.4
